# backend/internal/docs/swagger.yaml
basePath: /
definitions:
  internal_auth.AuthResponse:
    properties:
      access_token:
        type: string
      expires_at:
        type: string
      refresh_token:
        description: Future-ready placeholder; currently not issued.
        type: string
      token_type:
        type: string
      user:
        $ref: '#/definitions/internal_auth.UserDTO'
    type: object
  internal_auth.LoginRequest:
    properties:
      email:
        maxLength: 255
        type: string
      password:
        maxLength: 128
        minLength: 8
        type: string
    required:
    - email
    - password
    type: object
  internal_auth.RegisterRequest:
    properties:
      email:
        maxLength: 255
        type: string
      name:
        maxLength: 120
        minLength: 2
        type: string
      password:
        maxLength: 128
        minLength: 8
        type: string
    required:
    - email
    - name
    - password
    type: object
  internal_auth.UserDTO:
    properties:
      email:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
...
  skyport_internal_response.APIError:
    properties:
      code:
        type: string
      details: {}
      message:
        type: string
      request_id:
        type: string
    type: object
  skyport_internal_response.ErrorBody:
    properties:
      error:
        $ref: '#/definitions/skyport_internal_response.APIError'
    type: object
host: localhost
info:
  contact:
    name: Akash Halder
    url: https://github.com/Nil369
  description: Lightweight self-hosted developer cloud platform API
  license:
    name: AGPL-3.0
    url: https://www.gnu.org/licenses/agpl-3.0.en.html
  title: SkyPort API
  version: 0.0.1
paths:
...
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and return JWT access token
      parameters:
      - description: Login payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_auth.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_auth.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/skyport_internal_response.ErrorBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/skyport_internal_response.ErrorBody'
      summary: Login
      tags:
      - Auth
  /api/v1/auth/logout:
    post:
      description: Logout and revoke current token
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/skyport_internal_response.ErrorBody'
      security:
      - BearerAuth: []
      summary: Logout
      tags:
      - Auth
  /api/v1/auth/me:
    get:
      description: Returns details about the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_auth.UserDTO'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/skyport_internal_response.ErrorBody'
      security:
      - BearerAuth: []
      summary: Current user
      tags:
      - Auth
  /api/v1/auth/register:
    post:
      consumes:
      - application/json
      description: Register a new user account
      parameters:
      - description: Registration payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_auth.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_auth.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/skyport_internal_response.ErrorBody'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/skyport_internal_response.ErrorBody'
      summary: Register user
      tags:
      - Auth
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: 'Provide the JWT token as: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"